Abstract the Shelley ledger over `PParams`.
The set of protocol parameters will change in future eras. Some may be
added and, eventually, some removed. This commit paves the way for that
by making 'PParams' a type family.

Since we have to handle not only the set of parameters, but _updates_ to
the set of parameters, we additionally introduce a class `UsesPParams`
with an associated type `PParamsDelta`. Our requirement on this type is
minimal: we simply need a way to merge the delta into an existing set of
parameters. This is embodied in the function 'mergePPUpdates'.

We are now more explicit in each location as to which protocol
parameters we rely on.