slow down the proliferation of blue coins
In the Mary property tests, one of the types of generated multi-assets
has the behavior of (essentially) always using a new asset name.
These are the "blue coins". Unfortunately this has the problem of
clogging the UTxO, and in particular coin selection is not setup to
avoid creating multi-assets that are too big.

This change is a temporary band-aid to slow the proliferation of these
tokens, so that the nightly propertys tests will pass again. The better
solution is to avoid creating such large multi-assets.